refactor(server): extract incident/blast/runbook domain logic into packages/retrieval
Move pure analysis logic out of apps/server into well-named modules under
packages/retrieval/src/omniscience_retrieval/:

- incidents/resolution.py  (classification, confidence, response building)
- incidents/scoring.py     (calibrated scoring, DB persistence helpers)
- incidents/similar.py     (same-incident clustering, feature extraction)
- blast_radius.py          (impact scoring, path reconstruction, response)
- runbook.py               (step types, validators, scoring, event building)

apps/server modules become thin route→service orchestrators:
- incidents.py, incidents_scoring.py, blast_radius.py, runbook.py,
  similar_incidents.py re-export the domain surface and contain only
  FastAPI/app.state wiring, DB session loading, and Prometheus emission.

replay.py stays entirely in apps/server (pure dispatch wiring).
as_of.py stays in apps/server (Prometheus histogram + server telemetry).

Dependency DAG preserved: retrieval now declares omniscience-connectors,
omniscience-index, structlog, prometheus-client as explicit deps.
No behavior changes; all public API symbols remain importable from their
original paths via re-exports.
