+++ apps/server/src/omniscience_server/incidents.py
@@ -1,38 +1,10 @@
-"""resolve_incident composition (issue #153, Wave 2 of epic #99).
+"""resolve_incident orchestration layer (issue #153, Wave 2 of epic #99).
 
-This module is the design centre of the demo path: a single MCP/REST
-tool — :func:`resolve_incident` — that composes the four Wave-1
-incident sources (alerts, GitHub PR, Slack, OTel) into a recommendation
-bundle.  The connectors emit canonical-named entities and ``cross_ref``
-edges; the composition here is workspace-scoped, ACL-strict, bitemporal-
-aware, and does not invent any new retrieval primitives — it walks
-existing ``GraphStore.find_related`` output and classifies neighbours
-by canonical-name prefix.
-
-Composition shape
------------------
-
-1. **Parse** ``alert_id``: must be ``alert://{provider}/{provider_alert_id}``.
-2. **Resolve** the seed alert via :meth:`GraphStore.get_entity` — returns
-   ``404 alert_not_found`` when absent (cross-workspace alerts are
-   indistinguishable from "not found" by design, see #117).
-3. **Traverse** outward via :meth:`GraphStore.find_related` (max_depth
-   default 2).  Classify each related entity by canonical-name prefix:
-   - ``alert://`` -> the seed itself
-   - ``arn:`` / ``pod/`` / ``namespace/name`` / ``service://`` -> target resource
-   - ``trace://`` -> trace evidence (OTel #152)
-   - ``https://github.com/.../pull/N`` -> responsible PR (GitHub PR #150)
-   - ``slack://channel/{ch}/thread/{ts}`` -> Slack discussion (Slack #149)
-4. **Score** confidence with the v0.1 heuristic (see
-   :data:`CONFIDENCE_*` module-level constants).  This v0.1 placeholder
-   ships per the architect memo on epic #99; #155 replaces it with a
-   calibrated model.
-5. **Bundle** the recommendation into :class:`ResolveIncidentResponse`.
-6. **Cluster** past similar incidents (issue #233, additive field
-   ``similar_past``) via the same-incident clustering primitives in
-   :mod:`omniscience_server.similar_incidents`.  The clustering call is
-   best-effort: any failure returns an empty list rather than breaking
-   the primary resolve path.
+This module is the FastAPI-wiring layer for the ``resolve_incident`` MCP/REST
+tool.  Pure domain and analysis logic lives in
+:mod:`omniscience_retrieval.incidents`; this module composes the
+``GraphStore`` I/O, scoring-config loading, and ``similar_past``
+augmentation with the domain primitives.
 
 ACL invariant (non-negotiable)
 ------------------------------
@@ -41,234 +13,82 @@ ACL invariant (non-negotiable)
   input.  Every store call carries it (see :meth:`GraphStore` ACL
   contract from #117 / ADR-0005).
 - A foreign workspace's ``alert_id`` returns ``alert_not_found`` — the
-  same response a non-existent alert would produce.  Existence is
-  never leaked.
+  same response a non-existent alert would produce.
 
 ``as_of`` deferred from #173
 ----------------------------
 
-The forcing-function test ``test_resolve_incident_not_present_yet``
-(in ``tests/test_mcp_as_of.py``) is replaced with a positive contract
-test that asserts ``resolve_incident`` accepts ``as_of`` and forwards it
-to the bitemporal-aware ``GraphStore`` reads (#170 / #173 / #174).
-
-Workspace-mismatch HTTP status
-------------------------------
-
-Cross-workspace ``alert_id`` resolution returns ``404 alert_not_found``
-rather than ``403 forbidden`` to avoid leaking alert existence to a
-caller that cannot read the alert.  This mirrors the ``GraphStore``
-contract that "cross-workspace entities are indistinguishable from
-'not found' by design" (#117 / ADR-0005).  The trade-off is documented
-here so a future audit does not mis-classify it as an authz bug.
+See :mod:`omniscience_server.as_of` for bitemporal handling.
 """
 
 from __future__ import annotations
 
-import re
 import uuid
-from dataclasses import dataclass
 from datetime import datetime
-from typing import Any, Final
+from typing import Any
 
 import structlog
 from fastapi import FastAPI
-from omniscience_core.storage import EntityNodeView, GraphResultView, GraphStore
-from pydantic import BaseModel, Field, field_validator
+from omniscience_core.storage import GraphResultView, GraphStore
+
+# ---------------------------------------------------------------------------
+# Re-export the public API that external importers (tests, REST, MCP)
+# relied on from this module's original location.
+# ---------------------------------------------------------------------------
+from omniscience_retrieval.incidents.resolution import (
+    ALERT_NOT_FOUND_CODE,
+    CONFIDENCE_ALERT_ONLY,
+    CONFIDENCE_NONE,
+    CONFIDENCE_PR_NO_TEMPORAL_MATCH,
+    CONFIDENCE_PR_TEMPORAL_MATCH,
+    CONFIDENCE_RESOURCE_ONLY,
+    DEFAULT_MAX_DEPTH,
+    INVALID_ALERT_ID_CODE,
+    INVALID_ALERT_ID_MESSAGE,
+    MAX_MAX_DEPTH,
+    MAX_SLACK_THREADS_RETURNED,
+    MIN_MAX_DEPTH,
+    PR_RECENCY_WINDOW_SECONDS,
+    AlertIdRequest,
+    AlertSummary,
+    PrSummary,
+    ResolveIncidentResponse,
+    ResourceSummary,
+    SlackThreadSummary,
+)
+from omniscience_retrieval.incidents.resolution import (
+    build_meta as _build_meta,
+)
+from omniscience_retrieval.incidents.resolution import (
+    build_resolve_response as _build_response,
+)
+from omniscience_retrieval.incidents.resolution import (
+    classify_neighbours as _classify_neighbours,
+)
+from omniscience_retrieval.incidents.resolution import (
+    score_incident as _score_incident,
+)
+from omniscience_retrieval.incidents.resolution import (
+    validate_alert_id as _validate_alert_id,
+)
+from omniscience_retrieval.incidents.scoring import (
+    DEFAULT_CONFIDENCE_THRESHOLD,
+    IncidentScoringConfig,
+    load_workspace_config,
+)
 
 from omniscience_server.as_of import (
     enforce_utc,
     record_request_duration,
     resolve_effective_as_of,
 )
-from omniscience_server.incidents_scoring import (
-    DEFAULT_CONFIDENCE_THRESHOLD,
-    IncidentScoringConfig,
-    apply_weights,
-    compute_components,
-    load_workspace_config,
-)
 
 log = structlog.get_logger(__name__)
 
-# ---------------------------------------------------------------------------
-# Module constants — no magic numbers in the algorithm body
-# ---------------------------------------------------------------------------
-
-#: Default BFS depth when the caller does not override ``max_depth``.
-#: Issue #153 architect memo: depth 2 covers alert -> resource ->
-#: deploying-PR / matching Slack thread / observed trace; depth 3+ adds
-#: noise without recall.
-DEFAULT_MAX_DEPTH: Final[int] = 2
-
-#: Lower bound for ``max_depth``.  Depth < 1 makes the traversal a
-#: degenerate single-entity lookup which is :func:`mcp_get_entity`'s job.
-MIN_MAX_DEPTH: Final[int] = 1
-
-#: Upper bound for ``max_depth``.  Pinned to the same envelope as the
-#: GraphRAG composer's :data:`MAX_ANCHOR_DEPTH` so a future hub-entity
-#: blast does not surface here either (see #107).
-MAX_MAX_DEPTH: Final[int] = 5
-
-#: Soft cap on the number of Slack threads carried in the response.
-#: Bounds payload size; deeper traversals are still allowed but the
-#: response slices.
-MAX_SLACK_THREADS_RETURNED: Final[int] = 10
-
-#: Window inside which a PR merge is considered "temporally correlated"
-#: with an alert firing — issue #153 §C.  24h is the architect-memo
-#: cliff; tighter windows are punted to #155's calibrated model.
-PR_RECENCY_WINDOW_SECONDS: Final[int] = 24 * 60 * 60
-
-# ---------------------------------------------------------------------------
-# Confidence-score heuristic (v0.1 placeholder; #155 replaces it)
-# ---------------------------------------------------------------------------
-
-#: PR found AND merged within :data:`PR_RECENCY_WINDOW_SECONDS` before
-#: the alert fired.  Issue #153 §C.
-CONFIDENCE_PR_TEMPORAL_MATCH: Final[float] = 0.9
-
-#: PR found but no temporal correlation (older than the recency window
-#: or no merge timestamp on either side).  Issue #153 §C.
-CONFIDENCE_PR_NO_TEMPORAL_MATCH: Final[float] = 0.6
-
-#: Target resource resolved but no PR.  Issue #153 §C.
-CONFIDENCE_RESOURCE_ONLY: Final[float] = 0.4
-
-#: Only the alert entity itself was resolvable (no resource neighbour
-#: found in the traversal).  Issue #153 §C.
-CONFIDENCE_ALERT_ONLY: Final[float] = 0.1
-
-#: Alert resolution failed entirely.  Documented for completeness even
-#: though the surrounding handler 404s before reaching the heuristic.
-CONFIDENCE_NONE: Final[float] = 0.0
-
-# ---------------------------------------------------------------------------
-# Canonical-name prefix discriminators
-# ---------------------------------------------------------------------------
-
-_ALERT_PREFIX: Final[str] = "alert://"
-_TRACE_PREFIX: Final[str] = "trace://"
-_SLACK_THREAD_PREFIX: Final[str] = "slack://channel/"
-_SERVICE_PREFIX: Final[str] = "service://"
-_ARN_PREFIX: Final[str] = "arn:"
-_GITHUB_PR_RE: Final[re.Pattern[str]] = re.compile(
-    r"^https?://github\.com/[^/]+/[^/]+/pull/(\d+)(?:[/?#].*)?$"
-)
-_POD_RE: Final[re.Pattern[str]] = re.compile(r"^pod/[A-Za-z0-9._-]+$")
-# Generic ``<namespace>/<resource>`` k8s reference, e.g. ``default/api``.
-_K8S_NAMESPACED_RE: Final[re.Pattern[str]] = re.compile(r"^[A-Za-z0-9._-]+/[A-Za-z0-9._-]+$")
-
-# ---------------------------------------------------------------------------
-# Error codes (mirrored on the MCP / REST surfaces)
-# ---------------------------------------------------------------------------
-
-INVALID_ALERT_ID_CODE: Final[str] = "invalid_alert_id"
-INVALID_ALERT_ID_MESSAGE: Final[str] = (
-    "alert_id must be of the form 'alert://{provider}/{provider_alert_id}'"
-)
-ALERT_NOT_FOUND_CODE: Final[str] = "alert_not_found"
-
-
-# ---------------------------------------------------------------------------
-# Pydantic response schemas
-# ---------------------------------------------------------------------------
-
-
-class AlertSummary(BaseModel):
-    """Minimal alert payload returned in the recommendation bundle."""
-
-    name: str = Field(description="Canonical alert name (alert://provider/id).")
-    kind: str | None = None
-    source: str | None = None
-    chunk_text: str | None = None
-    valid_from: datetime | None = None
-
-
-class ResourceSummary(BaseModel):
-    """Minimal resource payload (pod / ARN / service / k8s name)."""
-
-    name: str = Field(description="Canonical resource identifier.")
-    kind: str | None = None
-    source: str | None = None
-    edge_type: str | None = Field(
-        default=None,
-        description="The edge_type by which the alert reached this resource.",
-    )
-
-
-class PrSummary(BaseModel):
-    """Minimal PR payload returned as the responsible PR (when any)."""
-
-    name: str = Field(description="Canonical PR URL.")
-    kind: str | None = None
-    source: str | None = None
-    edge_type: str | None = None
-    merged_at: datetime | None = Field(
-        default=None,
-        description=(
-            "Best-effort merge timestamp; v0.1 reads it from the entity's "
-            "bitemporal ``valid_from`` when populated by the connector "
-            "(GitHub PR #150).  Calibrated extraction lands in #155."
-        ),
-    )
-
-
-class SlackThreadSummary(BaseModel):
-    """Minimal Slack thread payload — name plus optional chunk text."""
-
-    name: str = Field(description="Canonical Slack thread URI.")
-    kind: str | None = None
-    source: str | None = None
-    chunk_text: str | None = None
-    edge_type: str | None = None
-
-
-class ResolveIncidentResponse(BaseModel):
-    """The recommendation bundle returned by ``resolve_incident``.
-
-    ``confidence_score`` is a v0.1 placeholder per the architect memo on
-    epic #99; the calibrated model lands in #155.  Callers should treat
-    the score as a stable schema slot, not a calibrated probability.
-
-    ``similar_past`` (issue #233, additive) carries the ranked list of
-    past incidents that the same-incident clustering primitive matched
-    for this seed.  Each entry follows the
-    :class:`omniscience_server.similar_incidents.SimilarIncident`
-    schema (serialised as a JSON dict for transport-agnostic delivery).
-    Empty list when the clustering call returned no matches or could
-    not run (best-effort, never breaks the primary bundle).
-    """
-
-    alert: AlertSummary
-    target_resource: ResourceSummary | None = None
-    responsible_pr: PrSummary | None = None
-    slack_threads: list[SlackThreadSummary] = Field(default_factory=list)
-    confidence_score: float = Field(ge=0.0, le=1.0)
-    effective_as_of: datetime
-    meta: dict[str, Any] | None = None
-    similar_past: list[dict[str, Any]] = Field(
-        default_factory=list,
-        description=(
-            "Ranked past incidents similar to this one (issue #233). "
-            "Shape matches SimilarIncident; empty when clustering yields "
-            "no matches or is unavailable."
-        ),
-    )
-
-
-# ---------------------------------------------------------------------------
-# Internal classification dataclasses
-# ---------------------------------------------------------------------------
-
-
-@dataclass(frozen=True, slots=True)
-class _Classified:
-    """Bucketed view of a single traversal neighbour."""
-
-    target_resource: EntityNodeView | None
-    responsible_pr: EntityNodeView | None
-    slack_threads: tuple[EntityNodeView, ...]
+# Public re-export so existing importers of
+# ``omniscience_server.incidents._validate_alert_id`` keep working.
+validate_alert_id = _validate_alert_id
+_ALERT_PREFIX = "alert://"
 
 
 # ---------------------------------------------------------------------------
@@ -287,22 +107,18 @@ async def mcp_resolve_incident(
 
     Mirrors the shape of :func:`mcp_get_related_entities` (issue #153
     requirement) — same ``workspace_id`` ACL invariant, same ``as_of``
-    plumbing, same telemetry histogram (``omniscience_request_duration_seconds``
-    with ``tool="resolve_incident"``).
+    plumbing, same telemetry histogram.
 
     Parameters
     ----------
     app:
-        FastAPI app exposing ``app.state.graph_store`` (Neo4j adapter
-        in production, in-memory fake in tests).
+        FastAPI app exposing ``app.state.graph_store``.
     alert_id:
         Canonical alert URI ``alert://{provider}/{provider_alert_id}``.
-        Other forms raise ``ValueError("invalid_alert_id:...")``.
     workspace_id:
         REQUIRED. Forwarded to every ``GraphStore`` call.
     as_of:
-        Optional ADR-0008 §5 bitemporal anchor.  Reuses
-        :func:`enforce_utc` and the surrounding handler's parser.
+        Optional ADR-0008 §5 bitemporal anchor.
     max_depth:
         BFS depth from the alert seed.  Clamped to
         ``[MIN_MAX_DEPTH, MAX_MAX_DEPTH]``; default ``DEFAULT_MAX_DEPTH``.
@@ -328,7 +144,6 @@ async def mcp_resolve_incident(
                 patcher.mark_pre_history()
             raise ValueError(f"{ALERT_NOT_FOUND_CODE}:{alert_id}")
 
-        # Traversal: alert -> related entities (resources, PRs, threads, traces).
         try:
             graph_result = await graph_store.find_related(
                 entity_name=alert_id,
@@ -337,9 +152,6 @@ async def mcp_resolve_incident(
                 max_depth=clamped_depth,
             )
         except ValueError as exc:
-            # Seed disappeared between the get_entity and find_related
-            # calls — race against re-ingestion.  Return the alert-only
-            # bundle rather than 404, mirroring the empty-graph contract.
             if str(exc).startswith("entity_not_found:"):
                 graph_result = GraphResultView(seed=seed, related=[], edges=[])
             else:
@@ -354,10 +166,6 @@ async def mcp_resolve_incident(
             config=scoring_config,
         )
         meta = _build_meta(confidence=confidence, config=scoring_config)
-        # Issue #233 — additive ``similar_past`` augmentation.  Called
-        # AFTER the primary classification so any failure is contained
-        # and does not break the bundle.  Late import avoids a circular
-        # dependency (similar_incidents -> incidents).
         similar_past = await _augment_with_similar_past(
             app=app,
             alert_id=alert_id,
@@ -369,216 +177,14 @@ async def mcp_resolve_incident(
             classified=classified,
             confidence=confidence,
             as_of=normalised_as_of,
+            effective_as_of=resolve_effective_as_of(normalised_as_of),
             meta=meta,
             similar_past=similar_past,
         ).model_dump(mode="json")
 
 
 # ---------------------------------------------------------------------------
-# Validation helpers
-# ---------------------------------------------------------------------------
-
-
-def _validate_alert_id(alert_id: str) -> None:
-    """Reject alert ids that are not of the form ``alert://provider/id``."""
-    if not isinstance(alert_id, str):  # pragma: no cover - typing barrier
-        raise ValueError(f"{INVALID_ALERT_ID_CODE}:{INVALID_ALERT_ID_MESSAGE}")
-    if not alert_id.startswith(_ALERT_PREFIX):
-        raise ValueError(f"{INVALID_ALERT_ID_CODE}:{INVALID_ALERT_ID_MESSAGE}")
-    suffix = alert_id[len(_ALERT_PREFIX) :]
-    if "/" not in suffix:
-        raise ValueError(f"{INVALID_ALERT_ID_CODE}:{INVALID_ALERT_ID_MESSAGE}")
-    provider, _, provider_alert_id = suffix.partition("/")
-    if not provider.strip() or not provider_alert_id.strip():
-        raise ValueError(f"{INVALID_ALERT_ID_CODE}:{INVALID_ALERT_ID_MESSAGE}")
-
-
-# ---------------------------------------------------------------------------
-# Classification — pure, easy to unit test
-# ---------------------------------------------------------------------------
-
-
-def _classify_neighbours(result: GraphResultView) -> _Classified:
-    """Bucket BFS neighbours by canonical-name prefix.
-
-    Resolves a single ``target_resource`` (closest-depth wins; ties
-    broken by traversal order), a single ``responsible_pr`` (most
-    recently merged wins, see :func:`_pick_responsible_pr`), and a
-    list of Slack threads (capped at ``MAX_SLACK_THREADS_RETURNED``).
-    """
-    target_candidates: list[EntityNodeView] = []
-    pr_candidates: list[EntityNodeView] = []
-    thread_candidates: list[EntityNodeView] = []
-    for node in result.related:
-        if _is_resource(node.name):
-            target_candidates.append(node)
-        elif _is_pr(node.name):
-            pr_candidates.append(node)
-        elif _is_slack_thread(node.name):
-            thread_candidates.append(node)
-        # trace://* and other neighbours are not surfaced in v0.1; #154
-        # demo fixture decides whether to promote them.
-    target_resource = _pick_target_resource(target_candidates)
-    responsible_pr = _pick_responsible_pr(pr_candidates)
-    threads = tuple(thread_candidates[:MAX_SLACK_THREADS_RETURNED])
-    return _Classified(
-        target_resource=target_resource,
-        responsible_pr=responsible_pr,
-        slack_threads=threads,
-    )
-
-
-def _is_pr(name: str) -> bool:
-    """Return ``True`` iff ``name`` is a GitHub PR canonical URL."""
-    return bool(_GITHUB_PR_RE.match(name))
-
-
-def _is_slack_thread(name: str) -> bool:
-    return name.startswith(_SLACK_THREAD_PREFIX)
-
-
-def _is_resource(name: str) -> bool:
-    """Return ``True`` for pod / ARN / service / k8s namespaced names.
-
-    Excludes alert / trace / Slack / PR canonical forms so the bucketing
-    is mutually exclusive at the prefix layer.
-    """
-    if name.startswith(_ALERT_PREFIX):
-        return False
-    if name.startswith(_TRACE_PREFIX):
-        return False
-    if name.startswith(_SLACK_THREAD_PREFIX):
-        return False
-    if _GITHUB_PR_RE.match(name):
-        return False
-    if name.startswith(_ARN_PREFIX):
-        return True
-    if name.startswith(_SERVICE_PREFIX):
-        return True
-    if _POD_RE.match(name):
-        return True
-    return bool(_K8S_NAMESPACED_RE.match(name))
-
-
-def _pick_target_resource(
-    candidates: list[EntityNodeView],
-) -> EntityNodeView | None:
-    """Closest-depth wins; ties broken by traversal order (input order)."""
-    if not candidates:
-        return None
-    return min(candidates, key=lambda n: n.depth)
-
-
-def _pick_responsible_pr(
-    candidates: list[EntityNodeView],
-) -> EntityNodeView | None:
-    """Most recently merged wins; falls back to closest-depth.
-
-    The connector (GitHub PR #150) carries the merge timestamp in the
-    entity's bitemporal ``valid_from`` — when present.  Entities without
-    a populated ``valid_from`` fall back to BFS depth so a single PR
-    candidate is still surfaceable.
-    """
-    if not candidates:
-        return None
-    with_merge = [c for c in candidates if c.valid_from is not None]
-    if with_merge:
-        return max(with_merge, key=lambda n: n.valid_from or datetime.min)
-    return min(candidates, key=lambda n: n.depth)
-
-
-# ---------------------------------------------------------------------------
-# Confidence-score heuristic
-# ---------------------------------------------------------------------------
-
-
-def _compute_confidence(
-    *,
-    alert: EntityNodeView,
-    classified: _Classified,
-) -> float:
-    """v0.1 deterministic placeholder per issue #153 §C.
-
-    The 4-rung ladder:
-
-    - 0.9 if ``responsible_pr`` is present AND merged within
-      :data:`PR_RECENCY_WINDOW_SECONDS` BEFORE the alert fired.
-    - 0.6 if ``responsible_pr`` is present but no temporal correlation.
-    - 0.4 if ``target_resource`` resolved but no PR.
-    - 0.1 if only the alert entity itself was resolvable.
-
-    The 0.0 rung is documented in :data:`CONFIDENCE_NONE` for parity
-    with the issue body but never returned here — alert-resolution
-    failure 404s upstream of this call.
-    """
-    if classified.responsible_pr is not None:
-        if _is_temporally_correlated(alert=alert, pr=classified.responsible_pr):
-            return CONFIDENCE_PR_TEMPORAL_MATCH
-        return CONFIDENCE_PR_NO_TEMPORAL_MATCH
-    if classified.target_resource is not None:
-        return CONFIDENCE_RESOURCE_ONLY
-    return CONFIDENCE_ALERT_ONLY
-
-
-def _is_temporally_correlated(
-    *,
-    alert: EntityNodeView,
-    pr: EntityNodeView,
-) -> bool:
-    """Return ``True`` iff PR merge time precedes alert by < window.
-
-    Both timestamps are read from the entity's bitemporal ``valid_from``;
-    the connectors populate them deterministically (#150 / #151).  When
-    either side is missing the heuristic defaults to ``False`` — the
-    "no temporal correlation" rung.
-    """
-    pr_merged_at = pr.valid_from
-    alert_fired_at = alert.valid_from
-    if pr_merged_at is None or alert_fired_at is None:
-        return False
-    if pr_merged_at > alert_fired_at:
-        # Merge AFTER fire is a regression-on-rollback or noise; not a
-        # cause.  Cliff-out per architect memo.
-        return False
-    delta = (alert_fired_at - pr_merged_at).total_seconds()
-    return 0 <= delta <= PR_RECENCY_WINDOW_SECONDS
-
-
-# ---------------------------------------------------------------------------
-# Response assembly
-# ---------------------------------------------------------------------------
-
-
-def _build_response(
-    *,
-    alert: EntityNodeView,
-    classified: _Classified,
-    confidence: float,
-    as_of: datetime | None,
-    meta: dict[str, Any] | None = None,
-    similar_past: list[dict[str, Any]] | None = None,
-) -> ResolveIncidentResponse:
-    """Produce the bounded :class:`ResolveIncidentResponse` payload."""
-    return ResolveIncidentResponse(
-        alert=AlertSummary(
-            name=alert.name,
-            kind=alert.kind,
-            source=alert.source,
-            chunk_text=alert.chunk_text,
-            valid_from=alert.valid_from,
-        ),
-        target_resource=_summarise_resource(classified.target_resource),
-        responsible_pr=_summarise_pr(classified.responsible_pr),
-        slack_threads=[_summarise_thread(t) for t in classified.slack_threads],
-        confidence_score=confidence,
-        effective_as_of=resolve_effective_as_of(as_of),
-        meta=meta,
-        similar_past=list(similar_past or []),
-    )
-
-
-# ---------------------------------------------------------------------------
-# Calibrated scoring + threshold flag (issue #155)
+# Scoring-config loading (requires DB session from app.state)
 # ---------------------------------------------------------------------------
 
 
@@ -588,9 +194,7 @@ async def _load_scoring_config(
     """Look up the per-tenant scoring config; return ``None`` to use v0.1.
 
     Tolerates a missing or unwired ``db_session_factory`` so unit tests
-    that only stub the graph store keep passing — the tests in
-    :mod:`tests.test_mcp_resolve_incident` plant an :class:`AsyncMock`
-    factory, but legacy fixtures may not.
+    that only stub the graph store keep passing.
     """
     factory = getattr(app.state, "db_session_factory", None)
     if factory is None:
@@ -599,96 +203,12 @@ async def _load_scoring_config(
         async with factory() as session:
             return await load_workspace_config(session, workspace_id)
     except Exception:
-        # A misconfigured DB session must NEVER break live incident
-        # resolution; fall back gracefully.  Log at warning so the issue
-        # surfaces in structured logs without propagating the error to callers.
         log.warning(
             "scoring_config_load_failed",
             workspace_id=str(workspace_id),
             exc_info=True,
         )
         return None
-
-
-def _score_incident(
-    *,
-    alert: EntityNodeView,
-    classified: _Classified,
-    max_depth: int,
-    config: IncidentScoringConfig | None,
-) -> float:
-    """Pick the calibrated path or fall back to the v0.1 ladder.
-
-    The legacy ladder is preserved verbatim when the workspace has no
-    configured weights — the #184 contract tests assert the four
-    boundary values (0.9 / 0.6 / 0.4 / 0.1) and must keep passing.
-    """
-    if config is None or config.weights is None:
-        return _compute_confidence(alert=alert, classified=classified)
-    components = compute_components(
-        alert_valid_from=alert.valid_from,
-        pr_valid_from=(
-            classified.responsible_pr.valid_from if classified.responsible_pr is not None else None
-        ),
-        has_pr=classified.responsible_pr is not None,
-        has_resource=classified.target_resource is not None,
-        resource_depth=(
-            classified.target_resource.depth if classified.target_resource is not None else 0
-        ),
-        thread_count=len(classified.slack_threads),
-        max_depth=max_depth,
-        pr_recency_window_seconds=PR_RECENCY_WINDOW_SECONDS,
-    )
-    return apply_weights(components, config.weights)
-
-
-def _build_meta(
-    *,
-    confidence: float,
-    config: IncidentScoringConfig | None,
-) -> dict[str, Any] | None:
-    """Populate ``meta.below_trust_threshold`` when the score is below threshold.
-
-    Returns ``None`` when the score clears the threshold so the response
-    shape stays close to v0.1 for high-confidence calls.
-    """
-    threshold = config.confidence_threshold if config is not None else DEFAULT_CONFIDENCE_THRESHOLD
-    if confidence < threshold:
-        return {"below_trust_threshold": True, "confidence_threshold": threshold}
-    return None
-
-
-def _summarise_resource(node: EntityNodeView | None) -> ResourceSummary | None:
-    if node is None:
-        return None
-    return ResourceSummary(
-        name=node.name,
-        kind=node.kind,
-        source=node.source,
-        edge_type=node.edge_type,
-    )
-
-
-def _summarise_pr(node: EntityNodeView | None) -> PrSummary | None:
-    if node is None:
-        return None
-    return PrSummary(
-        name=node.name,
-        kind=node.kind,
-        source=node.source,
-        edge_type=node.edge_type,
-        merged_at=node.valid_from,
-    )
-
-
-def _summarise_thread(node: EntityNodeView) -> SlackThreadSummary:
-    return SlackThreadSummary(
-        name=node.name,
-        kind=node.kind,
-        source=node.source,
-        chunk_text=node.chunk_text,
-        edge_type=node.edge_type,
-    )
 
 
 # ---------------------------------------------------------------------------
@@ -703,13 +223,7 @@ async def _augment_with_similar_past(
     workspace_id: uuid.UUID,
     as_of: datetime | None,
 ) -> list[dict[str, Any]]:
-    """Best-effort call to :func:`collect_similar_past`.
-
-    Wrapped here so the import is late (avoids the circular dep) and
-    so any failure is contained — clustering errors must NEVER fail
-    the primary ``resolve_incident`` bundle (additive-field contract
-    from the issue #233 spec).
-    """
+    """Best-effort call to :func:`collect_similar_past`."""
     try:
         from omniscience_server.similar_incidents import collect_similar_past
     except ImportError:  # pragma: no cover - defensive
@@ -722,28 +236,6 @@ async def _augment_with_similar_past(
     )
 
 
-# ---------------------------------------------------------------------------
-# Pydantic field validators (kept here so the contract is one import away)
-# ---------------------------------------------------------------------------
-
-
-class AlertIdRequest(BaseModel):
-    """Lightweight wrapper used by the REST surface to validate ``alert_id``.
-
-    The MCP surface uses string args directly; this model centralises
-    the prefix check for the REST handler so it can map ``ValueError``
-    to ``HTTP 400 invalid_alert_id``.
-    """
-
-    alert_id: str
-
-    @field_validator("alert_id")
-    @classmethod
-    def _validate(cls, value: str) -> str:
-        _validate_alert_id(value)
-        return value
-
-
 __all__ = [
     "ALERT_NOT_FOUND_CODE",
     "CONFIDENCE_ALERT_ONLY",
@@ -751,6 +243,7 @@ __all__ = [
     "CONFIDENCE_PR_NO_TEMPORAL_MATCH",
     "CONFIDENCE_PR_TEMPORAL_MATCH",
     "CONFIDENCE_RESOURCE_ONLY",
+    "DEFAULT_CONFIDENCE_THRESHOLD",
     "DEFAULT_MAX_DEPTH",
     "INVALID_ALERT_ID_CODE",
     "INVALID_ALERT_ID_MESSAGE",
@@ -760,9 +253,11 @@ __all__ = [
     "PR_RECENCY_WINDOW_SECONDS",
     "AlertIdRequest",
     "AlertSummary",
+    "IncidentScoringConfig",
     "PrSummary",
     "ResolveIncidentResponse",
     "ResourceSummary",
     "SlackThreadSummary",
     "mcp_resolve_incident",
+    "validate_alert_id",
 ]
