+++ packages/retrieval/pyproject.toml
@@ -5,15 +5,21 @@ description = "Hybrid retrieval service (vector + BM25 + RRF) for Omniscience"
 requires-python = ">=3.12"
 dependencies = [
     "omniscience-core",
+    "omniscience-connectors",
     "omniscience-embeddings",
+    "omniscience-index",
     "sqlalchemy[asyncio]>=2.0.30",
     "asyncpg>=0.29.0",
     "httpx>=0.27.0",
+    "prometheus-client>=0.20.0",
+    "structlog>=24.2.0",
 ]
 
 [tool.uv.sources]
 omniscience-core = { workspace = true }
+omniscience-connectors = { workspace = true }
 omniscience-embeddings = { workspace = true }
+omniscience-index = { workspace = true }
 
 [build-system]
 requires = ["hatchling"]
