+++ packages/retrieval/src/omniscience_retrieval/incidents/__init__.py
@@ -1,0 +1,112 @@
+"""Incident resolution domain logic for Omniscience.
+
+Re-exports the public surface from the submodules so callers can use:
+
+    from omniscience_retrieval.incidents import (
+        resolve_incident_bundle,
+        IncidentScoringConfig,
+        ...
+    )
+"""
+
+from .resolution import (
+    ALERT_NOT_FOUND_CODE,
+    CONFIDENCE_ALERT_ONLY,
+    CONFIDENCE_NONE,
+    CONFIDENCE_PR_NO_TEMPORAL_MATCH,
+    CONFIDENCE_PR_TEMPORAL_MATCH,
+    CONFIDENCE_RESOURCE_ONLY,
+    DEFAULT_MAX_DEPTH,
+    INVALID_ALERT_ID_CODE,
+    INVALID_ALERT_ID_MESSAGE,
+    MAX_MAX_DEPTH,
+    MAX_SLACK_THREADS_RETURNED,
+    MIN_MAX_DEPTH,
+    PR_RECENCY_WINDOW_SECONDS,
+    AlertSummary,
+    PrSummary,
+    ResolveIncidentResponse,
+    ResourceSummary,
+    SlackThreadSummary,
+    classify_neighbours,
+    compute_confidence,
+    pick_responsible_pr,
+    pick_target_resource,
+    score_incident,
+    validate_alert_id,
+)
+from .scoring import (
+    DEFAULT_CONFIDENCE_THRESHOLD,
+    DEFAULT_WEIGHTS,
+    SETTINGS_KEY,
+    WEIGHT_SUM_TOLERANCE,
+    IncidentScoringConfig,
+    IncidentScoringResponse,
+    IncidentScoringUpdateRequest,
+    IncidentScoringWeights,
+    apply_weights,
+    compute_components,
+    load_workspace_config,
+    save_workspace_config,
+)
+from .similar import (
+    MAX_LIMIT,
+    MAX_SINCE_DAYS,
+    MIN_LIMIT,
+    MIN_SINCE_DAYS,
+    FindSimilarIncidentsResponse,
+    SimilarIncident,
+    SimilarIncidentRequest,
+    build_features,
+    discover_candidates,
+    rank_similar_incidents,
+)
+
+__all__ = [
+    "ALERT_NOT_FOUND_CODE",
+    "CONFIDENCE_ALERT_ONLY",
+    "CONFIDENCE_NONE",
+    "CONFIDENCE_PR_NO_TEMPORAL_MATCH",
+    "CONFIDENCE_PR_TEMPORAL_MATCH",
+    "CONFIDENCE_RESOURCE_ONLY",
+    "DEFAULT_CONFIDENCE_THRESHOLD",
+    "DEFAULT_MAX_DEPTH",
+    "DEFAULT_WEIGHTS",
+    "INVALID_ALERT_ID_CODE",
+    "INVALID_ALERT_ID_MESSAGE",
+    "MAX_LIMIT",
+    "MAX_MAX_DEPTH",
+    "MAX_SINCE_DAYS",
+    "MAX_SLACK_THREADS_RETURNED",
+    "MIN_LIMIT",
+    "MIN_MAX_DEPTH",
+    "MIN_SINCE_DAYS",
+    "PR_RECENCY_WINDOW_SECONDS",
+    "SETTINGS_KEY",
+    "WEIGHT_SUM_TOLERANCE",
+    "AlertSummary",
+    "FindSimilarIncidentsResponse",
+    "IncidentScoringConfig",
+    "IncidentScoringResponse",
+    "IncidentScoringUpdateRequest",
+    "IncidentScoringWeights",
+    "PrSummary",
+    "ResolveIncidentResponse",
+    "ResourceSummary",
+    "SimilarIncident",
+    "SimilarIncidentRequest",
+    "SlackThreadSummary",
+    "apply_weights",
+    "build_features",
+    "classify_neighbours",
+    "compute_components",
+    "compute_confidence",
+    "discover_candidates",
+    "load_workspace_config",
+    "pick_responsible_pr",
+    "pick_target_resource",
+    "rank_similar_incidents",
+    "save_workspace_config",
+    "score_incident",
+    "validate_alert_id",
+]
